refact: reconcile encryption functions and standardize db encryption env variables
