+++ config/ci.json
@@ -31,9 +31,5 @@
   },
   "fixer": {
     "accessKey": "UNDEFINED_FIXER_ACCESS_KEY"
-  },
-  "twoFactorAuth": {
-    "secretKey": "guineapigs",
-    "cipher": "DES"
   }
 }
