+++ server/controllers/users.js
@@ -1,9 +1,9 @@
 import config from 'config';
-import crypto from 'crypto-js';
 import speakeasy from 'speakeasy';
 
 import * as auth from '../lib/auth';
 import emailLib from '../lib/email';
+import { crypto } from '../lib/encryption';
 import errors from '../lib/errors';
 import logger from '../lib/logger';
 import RateLimit, { ONE_HOUR_IN_SECONDS } from '../lib/rate-limit';
@@ -11,9 +11,6 @@ import { isValidEmail } from '../lib/utils';
 import models from '../models';
 
 const { Unauthorized } = errors;
-
-const SECRET_KEY = config.twoFactorAuth.secretKey;
-const CIPHER = config.twoFactorAuth.cipher;
 
 /**
  *
@@ -98,9 +95,7 @@ export const updateToken = async (req, res, next) => {
   // if there is a 2FA code we need to verify it before returning the token
   if (twoFactorAuthenticatorCode) {
     const encryptedTwoFactorAuthToken = req.remoteUser.twoFactorAuthToken;
-    const decryptedTwoFactorAuthToken = crypto[CIPHER].decrypt(encryptedTwoFactorAuthToken, SECRET_KEY).toString(
-      crypto.enc.Utf8,
-    );
+    const decryptedTwoFactorAuthToken = crypto.decrypt(encryptedTwoFactorAuthToken);
     const verified = speakeasy.totp.verify({
       secret: decryptedTwoFactorAuthToken,
       encoding: 'base32',
