+++ test/server/lib/encryption.test.js
@@ -1,21 +1,23 @@
 import { expect } from 'chai';
 
-import { decrypt, encrypt, generateKey } from '../../../server/lib/encryption';
+import { generateKey, secretbox } from '../../../server/lib/encryption';
 
 describe('server/lib/encryption', () => {
-  it('it encrypts and decrypts ok', () => {
-    const message = 'OpenCollective Rules';
-    const buff = Buffer.from(message);
-    const key = generateKey();
+  describe('secretbox', () => {
+    it('it encrypts and decrypts ok', () => {
+      const message = 'OpenCollective Rules';
+      const buff = Buffer.from(message);
+      const key = generateKey();
 
-    const encrypted = encrypt(buff, key);
+      const encrypted = secretbox.encrypt(buff, key);
 
-    expect(Buffer.isBuffer(encrypted)).to.be.true;
+      expect(Buffer.isBuffer(encrypted)).to.be.true;
 
-    expect(encrypted).to.not.eq(message);
+      expect(encrypted).to.not.eq(message);
 
-    const result = decrypt(encrypted, key);
+      const result = secretbox.decrypt(encrypted, key);
 
-    expect(result).to.eq(message);
+      expect(result).to.eq(message);
+    });
   });
 });
